Added caching hooks
A way to enable and implement caching for the package is now implemented.
Clients may use the Store type to make load operations consult a cache
(described by the Cache interface) before issuing requests to the Mojang
servers.

The LoadProperties and LoadNameHistory methods of Profile have been
updated to consult the cache of the Store their Profile was loaded
with in case their information not already is loaded.